requirements.txt
tensorflow==2.12.0
scipy>=1.10.0
numpy
matplotlib==3.10.1
pillow==11.2.1
streamlit==1.45.0

